+++ cosim_example_demos/TVB_NEST_demo/tvb_sim/utils_tvb.py
