+++ ms_manager.py
@@ -15,12 +15,21 @@ import os
 import json
 
 # Co-Simulator imports
-import common
-import configurations_manager
-from default_directories_enum import DefaultDirectories
-from launching_manager import LaunchingManager
+from common import args
+from EBRAINS_ConfigManager.workflow_configuraitons_manager.xml_parsers import enums
+from EBRAINS_ConfigManager.workflow_configuraitons_manager.xml_parsers import variables
+from EBRAINS_ConfigManager.workflow_configuraitons_manager.xml_parsers import plan_xml_manager
+from EBRAINS_ConfigManager.workflow_configuraitons_manager.xml_parsers import xml_tags
+from EBRAINS_ConfigManager.workflow_configuraitons_manager.xml_parsers import variables_manager
+from EBRAINS_ConfigManager.workflow_configuraitons_manager.xml_parsers import parameters_xml_manager
+from EBRAINS_ConfigManager.workflow_configuraitons_manager.xml_parsers import actions_xml_manager
+from EBRAINS_ConfigManager.workflow_configuraitons_manager.xml_parsers import arranger
+from EBRAINS_ConfigManager.global_configurations_manager.xml_parsers import configurations_manager
+from EBRAINS_ConfigManager.global_configurations_manager.xml_parsers.default_directories_enum import DefaultDirectories
+from EBRAINS_Launcher.launching_manager import LaunchingManager
 
-class CoSimulator:
+
+class MSManager:
     """
         Class representing the Co-Simulator tool
     Methods:
@@ -70,23 +79,23 @@ class CoSimulator:
 
         results_dir = self.__configuration_manager.get_directory(DefaultDirectories.RESULTS)
         json_output_filename = \
-            self.__parameters_parameters_for_json_file_dict[common.xml_tags.CO_SIM_XML_CO_SIM_PARAMS_FILENAME]
+            self.__parameters_parameters_for_json_file_dict[xml_tags.CO_SIM_XML_CO_SIM_PARAMS_FILENAME]
         json_output_path_filename = os.path.join(results_dir, json_output_filename)
 
         try:
             with open(json_output_path_filename, 'w') as json_output_file:
                 json.dump(
-                    self.__parameters_parameters_for_json_file_dict[common.xml_tags.CO_SIM_XML_CO_SIM_PARAMS_JSON_FILE],
+                    self.__parameters_parameters_for_json_file_dict[xml_tags.CO_SIM_XML_CO_SIM_PARAMS_JSON_FILE],
                     json_output_file)
         except OSError:
             self.__logger.error('{} cannot be created, OS error'.format(json_output_path_filename))
-            return common.enums.CoSimulatorReturnCodes.JSON_FILE_OS_ERROR
+            return enums.CoSimulatorReturnCodes.JSON_FILE_OS_ERROR
 
         self.__logger.info('Co-Simulation parameters were transformed successfully')
         self.__logger.info('Co-Simulation parameters: {}'.format(json_output_path_filename))
-        return common.enums.CoSimulatorReturnCodes.OK
+        return enums.CoSimulatorReturnCodes.OK
 
-    def run(self, args=None):
+    def run(self, args1=None):
         """
             Entry point of the Co-Simulation Co-Simulator tool
         :param args:
@@ -98,10 +107,10 @@ class CoSimulator:
         # STEP 1 - Checking command line parameters
         ########
         try:
-            self.__args = common.args.arg_parse()
+            self.__args = args.arg_parse()
         except SystemExit as e:
             # argument parser has reported some issue with the arguments
-            return common.enums.CoSimulatorReturnCodes.PARAMETER_ERROR
+            return enums.CoSimulatorReturnCodes.PARAMETER_ERROR
 
         ########
         # STEP 2 - Setting Up the Configuration Manager
@@ -132,12 +141,12 @@ class CoSimulator:
         ########
         self.__logger.info('Co-Simulator STEP 3 running')
         self.__variables_manager = \
-            common.variables_manager.VariablesManager(logger=self.__logger)
+            variables_manager.VariablesManager(logger=self.__logger)
 
         # STEP 3.1 - Setting Up the output location (path) for results
 #        self.__variables_manager.set_value(common.variables.CO_SIM_RESULTS_DIR,
  #                                          self.__configuration_manager.get_directory('results'))
-        self.__variables_manager.set_value(common.variables.CO_SIM_RESULTS_DIR,
+        self.__variables_manager.set_value(variables.CO_SIM_RESULTS_DIR,
                                            self.__configuration_manager.get_directory(DefaultDirectories.RESULTS))
 
 
@@ -145,20 +154,20 @@ class CoSimulator:
 
         self.__logger.info(
             f'Co-Simulator STEP 3 done, Co-Simulation results location: '
-            f'{self.__variables_manager.get_value(common.variables.CO_SIM_RESULTS_DIR)}')
+            f'{self.__variables_manager.get_value(variables.CO_SIM_RESULTS_DIR)}')
 
         ########
         # STEP 4 - Co-Simulation Plan
         ########
         self.__logger.info('Co-Simulator STEP 4, dissecting Co-Simulation Action Plan')
         self.__plan_xml_manager = \
-            common.plan_xml_manager.PlanXmlManager(configuration_manager=self.__configuration_manager,
+            plan_xml_manager.PlanXmlManager(configuration_manager=self.__configuration_manager,
                                                    logger=self.__logger,
                                                    xml_filename=self.__args.action_plan)
 
         # STEP 4.1 - Dissecting the Co-Simulation Plan XML file
-        if not self.__plan_xml_manager.dissect() == common.enums.XmlManagerReturnCodes.XML_OK:
-            return common.enums.CoSimulatorReturnCodes.XML_ERROR
+        if not self.__plan_xml_manager.dissect() == enums.XmlManagerReturnCodes.XML_OK:
+            return enums.CoSimulatorReturnCodes.XML_ERROR
 
         # Variables -> Could contain references to Environment variables, e.g. ${HOME}
         # STEP 4.2 - Getting the variables found on the Co-Simulation Plan XML file
@@ -166,10 +175,10 @@ class CoSimulator:
 
         # STEP 4.3 -    Validating the references to the CO_SIM_* variables
         #               by filling up the environment variables dictionary
-        if not common.enums.VariablesReturnCodes.VARIABLE_OK == \
+        if not enums.VariablesReturnCodes.VARIABLE_OK == \
                self.__variables_manager.set_co_sim_variable_values_from_variables_dict(
                    self.__action_plan_variables_dict):
-            return common.enums.CoSimulatorReturnCodes.VARIABLE_ERROR
+            return enums.CoSimulatorReturnCodes.VARIABLE_ERROR
 
         # Parameters -> Could contain references to CO_SIM_ variables and become new CO_SIM_ variables
         # STEP 4.4 - Getting the parameters found on the Co-Simulation Plan XML file
@@ -177,51 +186,51 @@ class CoSimulator:
 
         # STEP 4.5 -    Validating the references to the CO_SIM_* variables on the <parameters> sections
         #               by creating the new CO_SIM_* variables by means of the variables manager
-        if not common.enums.ParametersReturnCodes.PARAMETER_OK == \
+        if not enums.ParametersReturnCodes.PARAMETER_OK == \
                self.__variables_manager.create_variables_from_parameters_dict(self.__action_plan_parameters_dict):
-            return common.enums.CoSimulatorReturnCodes.PARAMETER_ERROR
+            return enums.CoSimulatorReturnCodes.PARAMETER_ERROR
 
         # STEP 4.6 - Creates Co-Simulation variables based on the information
         #            set on the variables and parameters sections of the processing XML action plan file
-        if not common.enums.VariablesReturnCodes.VARIABLE_OK == \
+        if not enums.VariablesReturnCodes.VARIABLE_OK == \
                self.__variables_manager.create_co_sim_run_time_variables(self.__action_plan_variables_dict,
                                                                          self.__action_plan_parameters_dict):
-            return common.enums.CoSimulatorReturnCodes.VARIABLE_ERROR
+            return enums.CoSimulatorReturnCodes.VARIABLE_ERROR
 
         # Action Plan -> ordered and grouped sequence of actions to achieve the Co-Simulation Experiment
         # STEP 4.7 - Getting the action plan per se
         self.__action_plan_dict = self.__plan_xml_manager.get_action_plan_dict()
 
-        self.__logger.info('{} -> {}'.format(common.variables.CO_SIM_ACTIONS_DIR,
-                                             self.__variables_manager.get_value(common.variables.CO_SIM_ACTIONS_DIR)))
-        self.__logger.info('{} -> {}'.format(common.variables.CO_SIM_ROUTINES_DIR,
-                                             self.__variables_manager.get_value(common.variables.CO_SIM_ROUTINES_DIR)))
+        self.__logger.info('{} -> {}'.format(variables.CO_SIM_ACTIONS_DIR,
+                                             self.__variables_manager.get_value(variables.CO_SIM_ACTIONS_DIR)))
+        self.__logger.info('{} -> {}'.format(variables.CO_SIM_ROUTINES_DIR,
+                                             self.__variables_manager.get_value(variables.CO_SIM_ROUTINES_DIR)))
         self.__logger.info('Co-Simulator STEP 4 done')
 
         ########
         # STEP 5 - Processing Co-Simulation Parameters
         ########
-        self.__logger.info('Co-Simulator STEP 5, dissecting Co-Simulation parameters')
-        self.__parameters_xml_manager = \
-            common.parameters_xml_manager.ParametersXmlManager(configuration_manager=self.__configuration_manager,
-                                                               logger=self.__logger,
-                                                               xml_filename=self.__args.parameters)
+        # self.__logger.info('Co-Simulator STEP 5, dissecting Co-Simulation parameters')
+        # self.__parameters_xml_manager = \
+        #     parameters_xml_manager.ParametersXmlManager(configuration_manager=self.__configuration_manager,
+        #                                                        logger=self.__logger,
+        #                                                        xml_filename=self.__args.parameters)
 
-        # STEP 5.1 - Dissecting Co-Simulation Parameters XML file
-        if not self.__parameters_xml_manager.dissect() == common.enums.XmlManagerReturnCodes.XML_OK:
-            return common.enums.CoSimulatorReturnCodes.XML_ERROR
+        # # STEP 5.1 - Dissecting Co-Simulation Parameters XML file
+        # if not self.__parameters_xml_manager.dissect() == enums.XmlManagerReturnCodes.XML_OK:
+        #     return enums.CoSimulatorReturnCodes.XML_ERROR
 
-        # STEP 5.2 - Getting the variables found in the Co-Simulation Parameters file
-        self.__parameters_variables_dict = self.__parameters_xml_manager.get_variables_dict()
+        # # STEP 5.2 - Getting the variables found in the Co-Simulation Parameters file
+        # self.__parameters_variables_dict = self.__parameters_xml_manager.get_variables_dict()
 
-        # STEP 5.3 - Getting the parameters found in the Co-Simulation Parameters file
-        self.__parameters_parameters_dict = self.__parameters_xml_manager.get_parameters_dict()
+        # # STEP 5.3 - Getting the parameters found in the Co-Simulation Parameters file
+        # self.__parameters_parameters_dict = self.__parameters_xml_manager.get_parameters_dict()
 
-        # STEP 5.4 - Getting the Co-Simulation parameters to be dump into a json file
-        self.__parameters_parameters_for_json_file_dict = self.__parameters_xml_manager.get_parameter_for_json_dict()
+        # # STEP 5.4 - Getting the Co-Simulation parameters to be dump into a json file
+        # self.__parameters_parameters_for_json_file_dict = self.__parameters_xml_manager.get_parameter_for_json_dict()
 
-        self.__logger.info('Co-Simulation parameters loaded from {}'.format(self.__args.parameters))
-        self.__logger.info('Co-Simulator STEP 5 done')
+        # self.__logger.info('Co-Simulation parameters loaded from {}'.format(self.__args.parameters))
+        # self.__logger.info('Co-Simulator STEP 5 done')
 
         ########
         # STEP 6 - Co-Simulation Actions (processing the XML configuration files)
@@ -229,12 +238,12 @@ class CoSimulator:
         self.__logger.info('Co-Simulator STEP 6, dissecting Co-Simulation Actions XML files')
         # STEP 6.1 - Getting the Actions Popen arguments, the CO_SIM_ variables transformation is performed 
         self.__actions_xml_manager = \
-            common.actions_xml_manager.ActionsXmlManager(configuration_manager=self.__configuration_manager,
+            actions_xml_manager.ActionsXmlManager(configuration_manager=self.__configuration_manager,
                                                          logger=self.__logger,
                                                          variables_manager=self.__variables_manager,
                                                          action_plan=self.__action_plan_dict)
-        if not self.__actions_xml_manager.dissect() == common.enums.XmlManagerReturnCodes.XML_OK:
-            return common.enums.CoSimulatorReturnCodes.XML_ERROR
+        if not self.__actions_xml_manager.dissect() == enums.XmlManagerReturnCodes.XML_OK:
+            return enums.CoSimulatorReturnCodes.XML_ERROR
 
         self.__actions_popen_args_dict = self.__actions_xml_manager.get_actions_popen_arguments_dict()
         self.__logger.info('Co-Simulator STEP 6 done')
@@ -245,22 +254,22 @@ class CoSimulator:
         self.__logger.info('Co-Simulator STEP 7, arranging environment')
         self.__items_to_be_arranged = self.__plan_xml_manager.get_items_to_be_arranged_dict()
 
-        self.__arranger = common.arranger.Arranger(configuration_manager=self.__configuration_manager,
+        self.__arranger = arranger.Arranger(configuration_manager=self.__configuration_manager,
                                                    logger=self.__logger,
                                                    variables_manager=self.__variables_manager,
                                                    items_to_be_arranged_dict=self.__items_to_be_arranged)
 
-        if not self.__arranger.arrange() == common.enums.ArrangerReturnCodes.OK:
-            return common.enums.CoSimulatorReturnCodes.ARRANGER_ERROR
+        if not self.__arranger.arrange() == enums.ArrangerReturnCodes.OK:
+            return enums.CoSimulatorReturnCodes.ARRANGER_ERROR
         self.__logger.info('Co-Simulator STEP 7 done')
 
         ########
         # STEP 8 - Converting Co-Simulation parameters from XML into JSON
         ########
-        self.__logger.info('Co-Simulator STEP 8, transforming Co-Simulation parameters')
-        if not self.generate_parameters_json_file() == common.enums.CoSimulatorReturnCodes.OK:
-            return common.enums.CoSimulatorReturnCodes.JSON_FILE_ERROR
-        self.__logger.info('Co-Simulator STEP 8 done')
+        # self.__logger.info('Co-Simulator STEP 8, transforming Co-Simulation parameters')
+        # if not self.generate_parameters_json_file() == enums.CoSimulatorReturnCodes.OK:
+        #     return enums.CoSimulatorReturnCodes.JSON_FILE_ERROR
+        # self.__logger.info('Co-Simulator STEP 8 done')
 
         ########
         # STEP 9 - Launching the Action Plan
@@ -275,10 +284,10 @@ class CoSimulator:
                                           self.__actions_popen_args_dict,
                                           self.__logger_settings,
                                           self.__configuration_manager)
-        if not launching_manager.carry_out_action_plan() == common.enums.LauncherReturnCodes.LAUNCHER_OK:
+        if not launching_manager.carry_out_action_plan() == enums.LauncherReturnCodes.LAUNCHER_OK:
             self.__logger.error('Error(s) were reported, check the errors log on {}'.format(
-                self.__variables_manager.get_value(common.variables.CO_SIM_RESULTS_DIR)))
-            return common.enums.CoSimulatorReturnCodes.LAUNCHER_ERROR
+                self.__variables_manager.get_value(variables.CO_SIM_RESULTS_DIR)))
+            return enums.CoSimulatorReturnCodes.LAUNCHER_ERROR
         # if not self.__launcher.carry_out_action_plan() == common.enums.LauncherReturnCodes.LAUNCHER_OK:
         #     self.__logger.error('Error(s) were reported, check the errors log on {}'.format(
         #         self.__variables_manager.get_value(common.variables.CO_SIM_RESULTS_DIR)))
@@ -289,7 +298,7 @@ class CoSimulator:
         # STEP 10 - Finishing
         ########
         self.__logger.info('Information about Co-Simulation process could be found on: {}'.format(
-            self.__variables_manager.get_value(common.variables.CO_SIM_RESULTS_DIR)))
+            self.__variables_manager.get_value(variables.CO_SIM_RESULTS_DIR)))
         self.__logger.info('END: Co-Simulation Co-Simulator')
 
-        return common.enums.CoSimulatorReturnCodes.OK
+        return enums.CoSimulatorReturnCodes.OK
